nesis/api/requirements.txt
argparse==1.4.0
bcrypt==4.0.1
dateparser==1.1.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-memcached==1.59
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.17.6
SQLAlchemy==2.0.25
StringGenerator==0.4.4
urllib3~=2.2.0
alembic==1.12.1
more-itertools==10.1.0
onnxruntime>=1.14.1
gevent==23.9.1
python-memcached==1.59
PyPDF2==3.0.1
smbprotocol==1.12.0
flask==3.0.0
boto3==1.34.75

Werkzeug>=3.0.3
dropbox~=11.36.2

# Document datasources
minio==7.2.0
Office365-REST-Python-Client==2.5.5
dropbox==11.36.2

injector==0.21.0

apscheduler==3.10.4
